update code structures
