+++ code/main.ts
@@ -1,15 +1,19 @@
-import BinarySearchTree from './src/tree/binary-search-tree';
+import BinarySearchTree from '@/data-structures/binary-search-tree';
 
-function getBSTRoot(insertArray: number[] = [1, 3, 5, 4, 2, 6]) {
-    const BST: BinarySearchTree = new BinarySearchTree();
-    insertArray.forEach(node => BST.insert(node));
-    const logNode = (key: any) => { }
-    BST.inOrderTraverse(logNode);
-    BST.preOrderTraverse(logNode);
-    BST.postOrderTraverse(logNode);
-    return BST.root;
+function getBSTRoot(insertArray: number[] = [1, 3, 5, 5, 4, 2, 6]) {
+  const BST: BinarySearchTree<number> = new BinarySearchTree();
+  insertArray.forEach((node) => BST.insert(node));
+  const logNode = (name: string, key: number) => {
+    console.log(name, key);
+  };
+  BST.inOrderTraverse(logNode.bind(null, 'inOrderTraverse'));
+  BST.preOrderTraverse(logNode.bind(null, 'preOrderTraverse'));
+  BST.postOrderTraverse(logNode.bind(null, 'postOrderTraverse'));
+  return BST.getRoot();
 }
 
-const Root = getBSTRoot();
-console.log(Root);
+const root = getBSTRoot();
 
+const rootEl = document.createElement('pre');
+rootEl.textContent = JSON.stringify(root, null, 4);
+document.querySelector('#app').appendChild(rootEl);
