+++ code/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "code",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "dependencies": {
+    "@types/lodash": "^4.14.157",
+    "html-webpack-plugin": "^4.3.0"
+  },
+  "devDependencies": {
+    "html-loader": "^1.1.0",
+    "ts-loader": "^7.0.5",
+    "typescript": "^3.9.5",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0"
+  },
+  "scripts": {
+    "build": "webpack --config build/webpack.config.js",
+    "dev": "webpack-dev-server --config build/webpack.config.js"
+  },
+  "author": "",
+  "license": "ISC"
+}
