+++ code/src/data-structures/double-linked-list.js
