--- code/src/models/node.ts
@@ -1,10 +1,0 @@
-export default class Node<K> {
-    left: Node<K>;
-    right: Node<K>;
-
-    constructor(public key: K) { }
-
-    toString() {
-        return `${this.key}`;
-    }
-}
