--- code/src/stack/stack-array.js
@@ -1,25 +1,0 @@
-class Stack {
-    constructor() {
-        this.items = []
-    }
-    push(element) {
-        this.items.push(element);
-    }
-    pop() {
-        return this.items.pop();
-    }
-    peek() {
-        return this.items[this.items.length - 1]
-    }
-    clear() {
-        this.items.length = 0;
-    }
-    isEmpty() {
-        return this.items.length === 0;
-    }
-    size() {
-        return this.items.length;
-    }
-}
-
-exports = module.exports = Stack
