+++ code/src/utils/helps.ts
@@ -1,0 +1,9 @@
+export const log = (exampleName: string, example_xxx: string): void => {
+  console.log(
+    `------------------------------------${exampleName} start------------------------------------`
+  );
+  console.log(example_xxx);
+  console.log(
+    `-------------------------------------${exampleName} end ------------------------------------`
+  );
+};
