--- code/src/utils/index.js
@@ -1,7 +1,0 @@
-const log = (exampleName, example_xxx) => {
-    console.log(`------------------------------------${exampleName} start------------------------------------`);
-    console.log(example_xxx);
-    console.log(`-------------------------------------${exampleName} end ------------------------------------`);
-}
-
-exports = module.exports = { log }
